+++ src/features/AgentTasks/AgentTaskDetail/TaskDetailPage.tsx
@@ -3,11 +3,10 @@ import { memo, useEffect } from 'react';
 
 import AutoSaveHint from '@/components/Editor/AutoSaveHint';
 import Loading from '@/components/Loading/BrandTextLoading';
-import DocumentModal from '@/features/DocumentModal';
+import DocumentPreviewModal from '@/features/DocumentModal/Preview';
 import NavHeader from '@/features/NavHeader';
 import ToggleRightPanelButton from '@/features/RightPanel/ToggleRightPanelButton';
 import WideScreenContainer from '@/features/WideScreenContainer';
-import { useDocumentStore } from '@/store/document';
 import { useGlobalStore } from '@/store/global';
 import { systemStatusSelectors } from '@/store/global/selectors';
 import { useTaskStore } from '@/store/task';
@@ -36,9 +35,6 @@ const TaskDetailPage = memo<TaskDetailPageProps>(({ taskId }) => {
   const useFetchTaskDetail = useTaskStore((s) => s.useFetchTaskDetail);
   const isLoading = useTaskStore(taskDetailSelectors.isTaskDetailLoading);
   const saveStatus = useTaskStore(taskDetailSelectors.taskSaveStatus);
-  const previewDocumentId = useDocumentStore((s) => s.previewDocumentId);
-  const closeDocumentPreview = useDocumentStore((s) => s.closeDocumentPreview);
-
   const [showTaskAgentPanel, toggleTaskAgentPanel] = useGlobalStore((s) => [
     systemStatusSelectors.showTaskAgentPanel(s),
     s.toggleTaskAgentPanel,
@@ -106,11 +102,7 @@ const TaskDetailPage = memo<TaskDetailPageProps>(({ taskId }) => {
         </WideScreenContainer>
       </Flexbox>
       <TopicChatDrawer />
-      <DocumentModal
-        documentId={previewDocumentId}
-        open={!!previewDocumentId}
-        onClose={closeDocumentPreview}
-      />
+      <DocumentPreviewModal />
     </Flexbox>
   );
 });
